requirements.txt
PyMuPDF==1.24.9
chromadb==0.4.4
sentence-transformers==2.2.2
openai==1.27.0
numpy==1.25.2
langchain==0.2.0
langchain-community==0.2.0
langchain-huggingface==0.0.5
langchain-chroma==0.0.5
langchain-ollama==0.0.5
langchain-text-splitters==0.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pypdf==4.2.0
python-dotenv==1.0.1
streamlit==1.36.0
groq==0.6.0
